# api-spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Future API",
    "description": "Endpoints for appointment scheduling.",
    "contact": {
      "name": "Faraz Fazli",
      "email": "farazfazli@gmail.com"
    },
    "version": "0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Localhost Development Backend"
    }
  ],
  "tags": [
    {
      "name": "Appointment Routes",
      "description": "All routes outlined in the design document."
    }
  ],
  "paths": {
    "/available-appointments": {
      "get": {
        "summary": "Get available appointments",
        "description": "Get all available appointments for a trainer between two dates.",
        "tags": [
          "Appointment Routes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "trainer_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Trainer's UUID",
            "example": "1391b756-cf01-4fd4-b0ae-eae8cfe2f46f"
          },
          {
            "in": "query",
            "name": "starts_at",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Starting date and time",
            "example": "2022-11-15T08:00:00-08:00"
          },
          {
            "in": "query",
            "name": "ends_at",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Ending date and time",
            "example": "2022-11-15T13:30:00-08:00"
          }
        ],
        "responses": {
          "200": {
            "description": "Available appointments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Available appointment time"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed data received"
          },
          "422": {
            "description": "Unable to fetch available appointments for that trainer"
          }
        }
      }
    },
    "/appointments": {
      "post": {
        "summary": "Post a new appointment",
        "description": "Add a new appointment for a user.",
        "tags": [
          "Appointment Routes"
        ],
        "requestBody": {
          "description": "Appointment details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "trainer_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Trainer's UUID"
                  },
                  "user_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Member's UUID"
                  },
                  "starts_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Starting date and time"
                  },
                  "ends_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Ending date and time"
                  }
                },
                "required": [
                  "trainer_id",
                  "user_id",
                  "starts_at",
                  "ends_at"
                ],
                "example": {
                  "trainer_id": "1391b756-cf01-4fd4-b0ae-eae8cfe2f46f",
                  "user_id": "2d793328-0c42-4f10-8475-a0c10ae3fdb7",
                  "starts_at": "2022-11-15T12:00:00-08:00",
                  "ends_at": "2022-11-15T12:30:00-08:00"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1,
                      "description": "Available appointment time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed data received"
          },
          "422": {
            "description": "Appointment time isn't valid"
          },
          "409": {
            "description": "Appointment already exists"
          }
        }
      }
    },
    "/scheduled-appointments": {
      "get": {
        "summary": "Get scheduled appointments",
        "description": "Get a list of scheduled appointments for a trainer.",
        "tags": [
          "Appointment Routes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "trainer_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Trainer's UUID",
            "example": "1391b756-cf01-4fd4-b0ae-eae8cfe2f46f"
          },
          {
            "in": "query",
            "name": "starts_at",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Starting date and time",
            "example": "2022-11-15T08:00:00-08:00"
          },
          {
            "in": "query",
            "name": "ends_at",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Ending date and time",
            "example": "2022-11-15T13:30:00-08:00"
          }
        ],
        "responses": {
          "200": {
            "description": "Scheduled appointments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "Appointment ID",
                        "minimum": 1
                      },
                      "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Member's UUID"
                      },
                      "starts_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Starting date and time"
                      },
                      "ends_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Ending date and time"
                      },
                      "tranier_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Trainer's UUID"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed data received"
          },
          "422": {
            "description": "Unable to fetch scheduled appointments for that trainer"
          }
        }
      }
    }
  }
}